Kód stránky pro registraci přesunut do třídy
